requirements.txt
fastapi
uvicorn[standard]
fastmcp
supabase
python-jose[cryptography]
load_dotenv